requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
Flask==2.1.2
Werkzeug==2.1.2
Pillow==9.2.0
torch==1.12.0+cpu
torchvision==0.13.0+cpu